requirements.txt
# EXISTENTES (déjalas como están o actualiza si gustas)
flask==3.0.3
requests==2.32.3
python-dotenv==1.0.1
google-generativeai==0.8.3
wasenderapi==0.3.3
httpx==0.27.2
pydantic==2.8.2
gunicorn==22.0.0

# NUEVAS PARA RAG
faiss-cpu==1.8.0
pypdf==4.3.1
numpy==1.26.4
